updated test generation utilieis

+++ tests/utilities/generate_data.sh
@@ -13,14 +13,34 @@ run_sqpoller () {
     sudo chown -R jpiet ${ansible_dir}
     echo "SUZIEQ"
     python3 ${suzieq_dir}/poller/sq-poller -i ${ansible_file} -n ${name}  >> ${verbose_log} 2>&1 &
-    sleep 120
+    RESULT_sq=$?
+    if (( ${RESULT_sq} > 0 )) ; then
+        return RESULT_sq
+    fi
+    sleep 180
     pkill -f sq-poller
     ps auxwww | grep poller
     echo "SUZIEQ poller done"
     sleep 15
     python3 ${suzieq_dir}/cli/suzieq-cli table show
-    python3 ${suzieq_dir}/cli/suzieq-cli device unique --columns=namespace
+    RESULT_sq=$?
+    if (( ${RESULT_sq} > 0 )) ; then
+        return RESULT_sq
+    fi
+    table=$(python3 ${suzieq_dir}/cli/suzieq-cli device unique --columns=namespace --namespace=${name})
+    echo ${table}
+    RESULT_sq=$?
+    if (( ${RESULT_sq} > 0 )) ; then
+        return RESULT_sq
+    fi
+    data=$(echo ${table} | grep 14)
+    echo "DATA " ${data}
+    if [ -z "$data" ] ; then
+       echo "Missing hosts " ${table}
+       exit 1
+    fi
 
+    return 0
 }
 
 run_scenario () {
@@ -34,22 +54,41 @@ run_scenario () {
     echo "DEPLOY RESULTS $?" >> ${log} 2>&1
     sleep 15 #on fast machines, not everything is all the way up without sleep
     sudo ansible-playbook ping.yml >> ${verbose_log} 2>&1
-    RESULT=$?
-    echo "PING RESULTS $RESULT" >> ${log} 2>&1
-    run_sqpoller ${topology} ${name} >> ${log} 2>&1
-
+    RESULT_sc=$?
+    echo "PING RESULTS $RESULT_sc" >> ${log} 2>&1
+    #return RESULT_sq
 }
 
 run_protos () {
     topology="$1"
     proto="$2"
+
     scenario="$3"
     name="$4"
+    tries=2
 
     echo ${proto} >> ${log} 2>&1
-    vagrant_up
     cd ${proto}
-    run_scenario ${topology} ${proto} ${scenario} ${name}
+    RESULT_sc=99
+    while (( ${RESULT_sc} > 0 )) && (( ${tries} > 0 )); do
+        vagrant_down
+        vagrant_up
+        echo ${RESULT_sc} ${tries}
+        run_scenario ${topology} ${proto} ${scenario} ${name}
+        tries=$(expr ${tries} -1)
+        echo "RESULT: " ${RESULT_sc} ${tries}
+    done
+    if (( ${RESULT_sc} > 0 )) ; then
+        echo "FAILED vagrant or ansible" >> ${log}
+        vagrant_down
+        exit 1
+    fi
+    run_sqpoller ${topology} ${name} >> ${log} 2>&1
+    if (( ${RESULT_sq} > 0 )) ; then
+        echo "FAILED sqpoller" >> ${log}
+        vagrant_down
+        exit 1
+    fi
     vagrant_down
     cd ..
 }
@@ -63,9 +102,8 @@ vagrant_down () {
 vagrant_up () {
     echo 'foo' >> ${verbose_log}
     time sudo vagrant up >> ${verbose_log} 2>&1
-    echo "VAGRANT UP" >> ${log} 2>&1
+    echo "VAGRANT UP %?" >> ${log} 2>&1
     sudo vagrant status >> ${log} 2>&1
-
 }
 
 log=`pwd`/log
